More precision in keeping zombiefied villager data
Now I can track with much more precision when a villager if zombified,
so the zombie villager gets info from the correct source.

Also fixed issues of the zombification not updatin the client, since the
packets could arrive after the entity joins the world. Added a sync
method for a redundancy check.

ServerInfoTracker fully implemented for the current features, including
auto-cleanup.